Move the arg parsing into its own file
